removed DS_MODALFRAME style from all dialogues to achieve compatibility with VS2005's MFC 8.0 merged rev 1360:1461       UUIDGen.{h,cpp} from trunk merged rev 1326:1461       ItemField.{h,cpp} from trunk merged rev 1321:1461       sha256.{h,cpp} from trunk merged rev 1327:1461       MyString.{h,cpp} from trunk merged rev 1221:1461       PWCharPool.h from trunk merged rev 1321:1461       Util.{h,cpp} from trunk merged rev 1292:1461       ItemData.{h,cpp} from trunk copied all missing corelib files from trunk, though they're not used so far. But this should simplifie merging in the future
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/branches/V1_92PPC@1462 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/corelib/PWCharPool.h
@@ -88,7 +88,7 @@ class CPasswordCharPool
   size_t m_x[NUMTYPES+1]; // spread lengths along X axis
   TCHAR *m_char_arrays[NUMTYPES];
 
-  int m_sumlengths; // sum of all selected chartypes
+  size_t m_sumlengths; // sum of all selected chartypes
 
   // Following state vars set by ctor, used by MakePassword()
   const UINT m_pwlen;
